Database Seeding Completed

+++ database/migrations/2022_07_16_054915_create_patients_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->enum('gender', ['Male', 'FeMale', 'Trans']);
             $table->enum('blood_group', ['A+', 'A-', 'B+', 'B-', 'O+', 'O-', 'AB+', 'AB-']);
             $table->date('dob');
-            $table->integer('contact_no')->unique();
+            $table->bigInteger('contact_no')->unique();
             $table->longText('description')->nullable();
             $table->timestamps();
         });
